Improved Interface and Removed File Lag
